fix: improve int tests and examples code
This commit contains the changes from re-generating
the service (including int tests and examples) using the
version of swagger.yaml from my draft PR in the
"projects" repo.

In the service code, you'll see some changes that reflect
the very latest changes made to the API definition.

And in the integration tests and examples, you'll find
changes resulting from the addition of links and examples
in the API definition.  This eliminates a lot of the manual
changes.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>